Move IsLocalUrl to shared UrlValidationHelper and fix Login page open redirect
- Extract IsLocalUrl into UrlValidationHelper.cs shared utility
- Add IsSafeRedirectUrl that validates URI format and locality
- Fix Login.razor.cs GetRedirectUrl() to validate ReturnUrl
- Remove inlined IsLocalUrl from ValidateTokenMiddleware
